// backend/src/api/client/ApiTypes.ts
export type StrengthSlug =
  | 'carefulness'
  | 'compassion'
  | 'courage'
  | 'creativity'
  | 'curiosity'
  | 'enthusiasm'
  | 'fairness'
  | 'forgiveness'
  | 'gratitude'
  | 'grit'
  | 'honesty'
  | 'hope'
  | 'humour'
  | 'judgement'
  | 'kindness'
  | 'leadership'
  | 'love'
  | 'loveOfBeauty'
  | 'loveOfLearning'
  | 'modesty'
  | 'perseverance'
  | 'perspective'
  | 'selfRegulation'
  | 'socialIntelligence'
  | 'spirituality'
  | 'teamwork';
// ...
export type HostSprint = {
  id: string;
  isCompleted: boolean;
  isEnded: boolean;
  isCodeActive: boolean;
  expectedStrengthCount: number;
  players: GroupGamePlayer[];
  sharedStrengths: Array<{
    strength: StrengthSlug;
    from: string;
    to: string;
  }>;
  updatedAt: string;
};
// ...
export type GroupGamePlayer = {
  id: string;
  nickname: string;
  color: string;
  avatar: string;
};
// ...
export type UpdateSprintParameters = {
  id: string;
};
export type UpdateSprintResponse = HostSprint;
export type UpdateSprintRequest = {
  isEnded?: boolean;
};
// ...
export type PatchPlayerSprintEvent = {
  isStarted: boolean;
  isCompleted: boolean;
  isEnded: boolean;
  players: string[];
  updatedAt: string;
};